Pulled out test helpers into separate file
