+++ x/gov/handler_test.go
@@ -24,81 +24,6 @@ var (
 	hCharlie   = weavetest.NewCondition().Address()
 )
 
-func decodeProposalOptions(raw []byte) (weave.Msg, error) {
-	model := ProposalOptions{}
-	err := model.Unmarshal(raw)
-	if err != nil {
-		return nil, errors.Wrap(err, "cannot parse data into Options struct")
-	}
-	return weave.ExtractMsgFromSum(model.Option)
-}
-
-// returns TextResolutionMsg, UpdateElectorateMsg, UpdateElectionRuleMsg
-func generateOptions(t *testing.T) ([]byte, []byte, []byte) {
-	t.Helper()
-
-	textOpts := &ProposalOptions{
-		Option: &ProposalOptions_Text{
-			Text: &TextResolutionMsg{
-				Metadata:   &weave.Metadata{Schema: 1},
-				Resolution: "CI must be green before merging",
-			},
-		},
-	}
-	textOption, err := textOpts.Marshal()
-	assert.Nil(t, err)
-
-	electorateOpts := &ProposalOptions{
-		Option: &ProposalOptions_Electorate{
-			Electorate: &UpdateElectorateMsg{
-				Metadata:     &weave.Metadata{Schema: 1},
-				ElectorateID: weavetest.SequenceID(1),
-				DiffElectors: []Elector{{
-					Address: hCharlie,
-					Weight:  22,
-				}},
-			},
-		},
-	}
-	electorateOption, err := electorateOpts.Marshal()
-	assert.Nil(t, err)
-
-	ruleOpts := &ProposalOptions{
-		Option: &ProposalOptions_Rule{
-			Rule: &UpdateElectionRuleMsg{
-				Metadata:          &weave.Metadata{Schema: 1},
-				ElectionRuleID:    weavetest.SequenceID(1),
-				VotingPeriodHours: 5,
-				Threshold: Fraction{
-					Numerator:   5,
-					Denominator: 8,
-				},
-			},
-		},
-	}
-	ruleOption, err := ruleOpts.Marshal()
-	assert.Nil(t, err)
-
-	return textOption, electorateOption, ruleOption
-}
-
-// returns decodable struct that fails Validate(), bytes that cannot decode
-func generateInvalidOptions(t *testing.T) ([]byte, []byte) {
-	t.Helper()
-
-	missingOpts := &ProposalOptions{
-		Option: &ProposalOptions_Text{
-			Text: &TextResolutionMsg{
-				Metadata: &weave.Metadata{Schema: 1},
-			},
-		},
-	}
-	missingOption, err := missingOpts.Marshal()
-	assert.Nil(t, err)
-
-	return missingOption, []byte("foobar")
-}
-
 func TestCreateTextProposal(t *testing.T) {
 	now := weave.AsUnixTime(time.Now())
 
@@ -1474,78 +1399,3 @@ func TestDeleteProposal(t *testing.T) {
 // 		})
 // 	}
 // }
-
-// ctxAwareMutator is a call back interface to modify the passed proposal for test setup
-type ctxAwareMutator func(weave.Context, *Proposal)
-
-func withTextProposal(t *testing.T, db store.KVStore, ctx weave.Context, mods ...ctxAwareMutator) *ProposalBucket {
-	t.Helper()
-	// setup electorate
-	withElectorate(t, db)
-	// setup election rules
-	withElectionRule(t, db)
-	textOptions, _, _ := generateOptions(t)
-
-	// adapter to call fixture mutator with context
-	ctxMods := make([]func(*Proposal), len(mods)+1)
-	// give a valid option here
-	// TODO: merge into proposalFixture
-	ctxMods[0] = func(p *Proposal) {
-		p.RawOption = textOptions
-	}
-	for i := 0; i < len(mods); i++ {
-		j := i
-		ctxMods[j+1] = func(p *Proposal) {
-			if mods[j] == nil {
-				return
-			}
-			mods[j](ctx, p)
-		}
-	}
-	pBucket := NewProposalBucket()
-	proposal := proposalFixture(hAlice, ctxMods...)
-
-	if _, err := pBucket.Create(db, &proposal); err != nil {
-		t.Fatalf("unexpected error: %+v", err)
-	}
-	return pBucket
-}
-
-func withElectorate(t *testing.T, db store.KVStore) *Electorate {
-	t.Helper()
-	electorate := &Electorate{
-		Metadata: &weave.Metadata{Schema: 1},
-		Title:    "fooo",
-		Admin:    hBobby,
-		Electors: []Elector{
-			{Address: hAlice, Weight: 1},
-			{Address: hBobby, Weight: 10},
-		},
-		TotalElectorateWeight: 11,
-	}
-	sortByAddress(electorate.Electors)
-	electorateBucket := NewElectorateBucket()
-
-	if _, err := electorateBucket.Create(db, electorate); err != nil {
-		t.Fatalf("unexpected error: %+v", err)
-	}
-	return electorate
-}
-
-func withElectionRule(t *testing.T, db store.KVStore) *ElectionRule {
-	t.Helper()
-	rulesBucket := NewElectionRulesBucket()
-	rule := &ElectionRule{
-		Metadata:          &weave.Metadata{Schema: 1},
-		Title:             "barr",
-		Admin:             hBobby,
-		VotingPeriodHours: 1,
-		Threshold:         Fraction{1, 2},
-		ElectorateID:      weavetest.SequenceID(1),
-	}
-
-	if _, err := rulesBucket.Create(db, rule); err != nil {
-		t.Fatalf("unexpected error: %+v", err)
-	}
-	return rule
-}
